Move non-chart code to better locations
