+++ lib/dashboard/alerts/common/convert_kwh.rb
