+++ lib/dashboard/charting_and_reports/equivalences/energy_equivalences_conversions.rb
@@ -1,4 +1,4 @@
-require_relative '../scalar/format_energy_unit.rb'
+require_relative '../../utilities/format_energy_unit.rb'
 class EnergyEquivalences
   def self.format_unit(value, unit)
     case unit
