+++ lib/dashboard/utilities/format_energy_unit.rb
